refactor: collapse foundation modules into hl7v2
