+++ crates/hl7v2-escape/Cargo.toml
@@ -2,6 +2,7 @@
 readme = "README.md"
 name = "hl7v2-escape"
 version.workspace = true
+publish = false
 edition.workspace = true
 rust-version.workspace = true
 authors.workspace = true
@@ -13,7 +14,7 @@ keywords = ["hl7", "healthcare", "escape", "encoding"]
 categories = ["encoding", "parser-implementations"]
 
 [dependencies]
-hl7v2-model = { version = "1.2.0", path = "../hl7v2-model" }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false }
 
 [dev-dependencies]
 hl7v2-test-utils = { path = "../hl7v2-test-utils" }
