+++ crates/hl7v2/Cargo.toml
@@ -25,7 +25,6 @@ json = []
 profile = ["profile-core", "persistent-cache"]
 profile-core = [
     "dep:chrono",
-    "dep:serde",
     "dep:serde_yaml",
     "dep:regex",
     "dep:tokio",
@@ -41,7 +40,6 @@ stream = ["dep:tokio"]
 network = ["stream", "dep:tokio-util", "dep:futures", "dep:bytes"]
 synthetic = [
     "ack",
-    "dep:serde",
     "dep:chrono",
     "dep:rand",
     "dep:rand_distr",
@@ -50,15 +48,12 @@ synthetic = [
     "dep:uuid",
 ]
 redact = []
-lifecycle = ["dep:chrono", "dep:serde", "dep:sha2"]
-experimental-guard = ["dep:serde"]
+lifecycle = ["dep:chrono", "dep:sha2"]
+experimental-guard = []
 
 [dependencies]
-hl7v2-model = { version = "1.2.0", path = "../hl7v2-model" }
-hl7v2-escape = { version = "1.2.0", path = "../hl7v2-escape" }
-hl7v2-mllp = { version = "1.2.0", path = "../hl7v2-mllp" }
 thiserror = { workspace = true }
-serde = { workspace = true, features = ["derive"], optional = true }
+serde = { workspace = true, features = ["derive"] }
 chrono = { workspace = true, features = ["serde"], optional = true }
 sha2 = { workspace = true, optional = true }
 regex = { workspace = true, optional = true }
