+++ crates/hl7v2/src/transport/mod.rs
@@ -1,0 +1,6 @@
+//! HL7 v2 transport helpers.
+
+pub mod mllp;
+
+#[cfg(feature = "network")]
+pub mod network;
